refactor(runtime): restructure control serialization code
prep for RSC support
