--- packages/runtime/src/builder/serialization/index.ts
@@ -1,3 +1,0 @@
-export * from './control-serialization'
-export * from './function-serialization'
-export type { Serialize, Deserialize } from './controls/types'
