--- packages/runtime/src/controls/visitors/message-port-serializer/index.ts
@@ -1,6 +1,0 @@
-export { ClientMessagePortSerializationVisitor } from './client-message-port-serialization-visitor'
-export {
-  serializeFunction,
-  deserializeFunction,
-  functionDeserializationPlugin,
-} from './function-serialization'
